상세정보창에 디버프 창 생성
상세정보창에 디버프 창을 5개 생성, 스크립트와 연결했다.

Panel 스프라이트의 크기 때문에 밑에 여백이 생기는 문제가 있었다. 이에 None으로 변경하고, Outline을 컴포넌트로 추가했다.

그리고 Enemy를 프리팹화, 2개의 적을 연동한다.

Awake에 넣었던 IncreaseEnemy를 Start로 옮겼다. BattleInfo의 싱글톤과 충돌했기 때문

디버프 창을 스크립트로 연동했다. Player는 BehaviourPanel이 없어서, Dummy를 두었다.

Player에도 Status Panel과 Status Opener를 넣어줬다.

디버프 내용 갱신을 위한, DebuffInfo 스크립트를 추가했다. 내친 김에 Info를 모은 오브젝트 이름도 Battle Info로 변경, 안 쓰던 BattleManager는 삭제, Info들도 폴더에 묶었다.

Character의 Bleed, Debuff 관련 함수들을 간단히 리팩토링 했다.
