--- Assets/02. Scripts/Battles/BattleManager.cs
@@ -1,11 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using System.Runtime.CompilerServices;
-using UnityEngine;
-using UnityEngine.Events;
-
-// 턴 관리, 게임 오버, 게임 중단 등을 다루는 클래스
-public class BattleManager : MonoBehaviour
-{ 
-
-}
