--- Assets/02. Scripts/Battles/BattleManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 527187a62af881e45bccfe0097603b43
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
