--- Assets/02. Scripts/Battles/CardInfo.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 285625b4f776cb5458f4e9d6c9edf36c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
