--- Assets/02. Scripts/Battles/EnemySkillInfo.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c7553c1be591b4d42bce74038ae81ded
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
