+++ Assets/02. Scripts/Battles/Infomations.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 39da9ced8424b9b44aca69df8f268ce4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
