+++ Assets/02. Scripts/Battles/Infomations/BattleInfo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d24acc541c41a274fb4af241d3420fa0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
