+++ Assets/02. Scripts/Battles/Infomations/DebuffInfo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fb82421752845e54cb537ab06dc0f08f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
