+++ Assets/03. Prefebs/Enemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: adae741350cc13c4e9e18e84c40244bf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
